Split up the schemas
